// src/types/request.d.ts
import moment from 'moment';

export interface BindValiables {
    startTime: string;
    ddlStartYear: string;
    ddlStartMonth: string;
    ddlStartDay: string;
    ddlStartHour: string;
    ddlEndYear: string;
    ddlEndMonth: string;
    ddlEndDay: string;
    ddlEndHour: string;
    ddlArea: string;
    checkBoxMstList: string;
    download: string;
}